requirements.txt
fastapi
uvicorn
pydantic
requests
python-dotenv
google-generativeai
pdf2image
python-multipart
Pillow
